+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2024-02-21
+### Added
+- Stripe: it will take your money
+  - happy path only but it works, which is important
+
 ## [0.2.7] - 2024-02-21
 ### Updated
 - Updated the cart count to account for quantities of the same item